pyproject.toml
[project]
name = "doc-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.3.2",
    "faiss-cpu>=1.11.0.post1",
    "httpx>=0.28.1",
    "numpy>=2.3.2",
    "pdf2image>=1.17.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.55.0",
]
